Improve RDA: Clean code
  * avoid uneccessary shorteners: use readable names
  * don't clump files: extract some data structure in their own file
  * avoid code duplication
