+++ angr/knowledge_plugins/key_definitions/definition.py
@@ -2,52 +2,7 @@ from ...engines.light import SpOffset
 from ...code_location import CodeLocation
 from .atoms import Atom, MemoryLocation, Register
 from .dataset import DataSet
-
-
-class Tag:
-    """
-    A tag for a Definition that can carry different kinds of metadata.
-    """
-    def __repr__(self):
-        raise NotImplementedError()
-
-
-class ParamTag(Tag):
-    """
-    A tag for a definition of a parameter.
-    """
-
-    def __init__(self, metadata: object=None):
-        super(ParamTag, self).__init__()
-        self.metadata = metadata
-
-    def __repr__(self):
-        return '<ParamTag {Metadata:%s}>' % (self.metadata)
-
-
-class RetValueTag(Tag):
-    """
-    A tag for a definiton of a return value
-    of a function.
-    """
-    def __init__(self, metadata: object=None):
-        super(RetValueTag, self).__init__()
-        self.metadata = metadata
-
-    def __repr__(self):
-        return '<RetValueTag {Metadata:%s}>' % (self.metadata)
-
-
-class InitValueTag(Tag):
-    """
-    A tag for a definiton of an initial value
-    """
-    def __init__(self, metadata: object=None):
-        super(InitValueTag, self).__init__()
-        self.metadata = metadata
-
-    def __repr__(self):
-        return '<InitValueTag {Metadata:%s}>' % (self.metadata)
+from .tag import Tag
 
 
 class Definition:
